requirements.txt
# ── Core data ─────────────────────────────────────────────────────────────────
yfinance>=0.2.36
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
lxml>=5.0.0

# ── Environment ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Web UI ─────────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── Excel export ───────────────────────────────────────────────────────────────
openpyxl>=3.1.0

# ── Utilities ──────────────────────────────────────────────────────────────────
rich>=13.0.0
